+++ test/aeso_compiler_tests.erl
@@ -265,7 +265,9 @@ warnings() ->
         "The function `called_unused_function2` is defined but never used.">>,
       <<?PosW(48, 5)
         "Unused return value.">>,
-      <<?PosW(60, 5)
+      <<?PosW(53, 44)
+        "The constraint on the type variable `'a` is a duplication of the constraint at line 53, column 34">>,
+      <<?PosW(65, 5)
         "The function `dec` is defined but never used.">>
      ]).
 
@@ -805,6 +807,90 @@ failing_contracts() ->
                      "to arguments\n"
                      "  `1 : int`">>
                   ])
+    , ?TYPE_ERROR(comparable_typevar_constraints,
+                  [<<?Pos(21,30)
+                     "Values of type `'a` are not comparable by equality">>,
+                   <<?Pos(25,38)
+                     "The type variable `'b` is constrained but never used">>,
+                   <<?Pos(29,41)
+                     "Unknown constraint `foo` used on the type variable `'a`">>,
+                   <<?Pos(60,56)
+                     "Values of type `address` are not comparable by inequality">>,
+                   <<?Pos(63,58)
+                     "Values of type `Chain.ttl` are not comparable by inequality">>,
+                   <<?Pos(66,45)
+                     "Values of type `A` are not comparable by inequality">>,
+                   <<?Pos(73,47)
+                     "Values of type `(int, char) => bool` are not comparable by inequality">>,
+                   <<?Pos(74,47)
+                     "Values of type `(int, char) => bool` are not comparable by equality">>,
+                   <<?Pos(89,71)
+                     "Values of type `list(address)` are not comparable by inequality">>,
+                   <<?Pos(92,77)
+                     "Values of type `option(address)` are not comparable by inequality">>,
+                   <<?Pos(95,76)
+                     "Values of type `(address * int)` are not comparable by inequality">>,
+                   <<?Pos(96,76)
+                     "Values of type `(address * int)` are not comparable by equality">>,
+                   <<?Pos(100,68)
+                     "Values of type `list((int, char) => bool)` are not comparable by inequality">>,
+                   <<?Pos(101,68)
+                     "Values of type `list((int, char) => bool)` are not comparable by equality">>,
+                   <<?Pos(103,74)
+                     "Values of type `option((int, char) => bool)` are not comparable by inequality">>,
+                   <<?Pos(104,74)
+                     "Values of type `option((int, char) => bool)` are not comparable by equality">>,
+                   <<?Pos(106,73)
+                     "Values of type `((int, char) => bool * int)` are not comparable by inequality">>,
+                   <<?Pos(107,73)
+                     "Values of type `((int, char) => bool * int)` are not comparable by equality">>,
+                   <<?Pos(111,71)
+                     "Values of type `map(int, int)` are not comparable by inequality">>,
+                   <<?Pos(114,80)
+                     "Values of type `oracle(int, int)` are not comparable by inequality">>,
+                   <<?Pos(117,98)
+                     "Values of type `oracle_query(int, int)` are not comparable by inequality">>,
+                   <<?Pos(120,90)
+                     "Values of type `custom_datatype(int)` are not comparable by inequality">>,
+                   <<?Pos(123,84)
+                     "Values of type `custom_record(int)` are not comparable by inequality">>,
+                   <<?Pos(128,86)
+                     "Values of type `map(address, address)` are not comparable by inequality">>,
+                   <<?Pos(131,95)
+                     "Values of type `oracle(address, address)` are not comparable by inequality">>,
+                   <<?Pos(134,113)
+                     "Values of type `oracle_query(address, address)` are not comparable by inequality">>,
+                   <<?Pos(137,97)
+                     "Values of type `custom_datatype(address)` are not comparable by inequality">>,
+                   <<?Pos(140,91)
+                     "Values of type `custom_record(address)` are not comparable by inequality">>,
+                   <<?Pos(145,75)
+                     "Values of type `map((int, char) => bool, (int, char) => bool)` are not comparable by inequality">>,
+                   <<?Pos(146,75)
+                     "Values of type `map((int, char) => bool, (int, char) => bool)` are not comparable by equality">>,
+                   <<?Pos(148,84)
+                     "Values of type `oracle((int, char) => bool, (int, char) => bool)` are not comparable by inequality">>,
+                   <<?Pos(149,84)
+                     "Values of type `oracle((int, char) => bool, (int, char) => bool)` are not comparable by equality">>,
+                   <<?Pos(151,102)
+                     "Values of type `oracle_query((int, char) => bool, (int, char) => bool)` are not comparable by inequality">>,
+                   <<?Pos(152,102)
+                     "Values of type `oracle_query((int, char) => bool, (int, char) => bool)` are not comparable by equality">>,
+                   <<?Pos(154,94)
+                     "Values of type `custom_datatype((int, char) => bool)` are not comparable by inequality">>,
+                   <<?Pos(155,94)
+                     "Values of type `custom_datatype((int, char) => bool)` are not comparable by equality">>,
+                   <<?Pos(157,88)
+                     "Values of type `custom_record((int, char) => bool)` are not comparable by inequality">>,
+                   <<?Pos(158,88)
+                     "Values of type `custom_record((int, char) => bool)` are not comparable by equality">>,
+                   <<?Pos(162,35)
+                     "Values of type `map(int, int)` are not comparable by inequality">>,
+                   <<?Pos(163,35)
+                     "Values of type `('a) => 'a` are not comparable by inequality">>,
+                   <<?Pos(167,34)
+                     "Values of type `('b) => 'b` are not comparable by equality">>
+                  ])
     , ?TYPE_ERROR(warnings,
                   [<<?Pos(0, 0)
                       "The file `Triple.aes` is included but not used.">>,
@@ -834,7 +920,9 @@ failing_contracts() ->
                      "The function `called_unused_function2` is defined but never used.">>,
                    <<?Pos(48, 5)
                      "Unused return value.">>,
-                   <<?Pos(60, 5)
+                   <<?Pos(53, 44)
+                     "The constraint on the type variable `'a` is a duplication of the constraint at line 53, column 34">>,
+                   <<?Pos(65, 5)
                      "The function `dec` is defined but never used.">>
                   ])
     ].